# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin panel API",
        "description": "Codeyad-React Learning-Ecommerce admin panel API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/guarantees": {
            "get": {
                "tags": [
                    "Guarantees"
                ],
                "summary": "Get all guarantees",
                "description": "Get all guarantees",
                "operationId": "allGuarantees",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{...}, {...}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Guarantees"
                ],
                "summary": "Add Guarantee",
                "description": "Store one guarantee",
                "operationId": "addGuarantee",
                "requestBody": {
                    "description": "Add one guarantee",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "test guarantee"
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "some description about this record..."
                                    },
                                    "length": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "length_unit": {
                                        "type": "string",
                                        "example": "year"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/guarantees/{id}": {
            "get": {
                "tags": [
                    "Guarantees"
                ],
                "summary": "Get one guarantee",
                "description": "Get one guarantee with id",
                "operationId": "oneGuarantee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "{...}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Guarantees"
                ],
                "summary": "Edit guarantee",
                "description": "Edit one guarantee",
                "operationId": "editGuarantee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Edit guarantee",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "descriptions": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "number"
                                    },
                                    "length_unit": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "title": "Edited title",
                                "descriptions": "Edited description",
                                "length": "10",
                                "length_unit": "month"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "{...}"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ویرایش با موفقیت انجام شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Guarantees"
                ],
                "summary": "Delete guarantee",
                "description": "Delete one guarantee",
                "operationId": "deleteGuarantee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "حذف با موفقیت انجام شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}